requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
asgiref==3.7.2
drf-yasg==1.21.5
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0